requirements.txt
asgiref==3.3.1
Django==3.1.3
django-queryset-csv==1.1.0
joblib==1.0.0
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
unicodecsv==0.14.1
